Test mail service replaced with MockBean to fix the app lifecycle
